+++ assets/javascripts/kitten/components/form/comment.js
@@ -1,16 +1,17 @@
 import React, { Component, Fragment } from 'react'
 import Radium, { StyleRoot } from 'radium'
 import PropTypes from 'prop-types'
-import { Marger as MargerBase } from 'kitten/components/layout/marger'
+import { Marger } from 'kitten/components/layout/marger'
 import { CommentAvatar } from 'kitten/components/form/comment-avatar'
-import { Text } from 'kitten/components/typography/text'
+import { Text as TextBase } from 'kitten/components/typography/text'
 import { ScreenConfig } from 'kitten/constants/screen-config'
+import { mediaQueries } from 'kitten/hoc/media-queries'
 import COLORS from 'kitten/constants/colors-config'
 
-const Marger = Radium(MargerBase)
+const Text = Radium(TextBase)
 
-export class Comment extends Component {
-  static PropTypes = {
+class CommentComponent extends Component {
+  static propTypes = {
     text: PropTypes.string.isRequired,
     ownerName: PropTypes.string.isRequired,
     avatarImgProps: PropTypes.object.isRequired,
@@ -18,34 +19,42 @@ export class Comment extends Component {
   }
 
   render() {
-    const { avatarProps, commentDate } = this.props
+    const { avatarImgProps, commentDate } = this.props
 
     return (
-      <StyleRoot>
+      <Fragment>
         <div style={styles.grid}>
           <CommentAvatar
-            avatarImgProps={avatarProps}
+            avatarImgProps={avatarImgProps}
             commentDate={commentDate}
           />
           {this.renderComment()}
         </div>
-      </StyleRoot>
+      </Fragment>
     )
   }
 
   renderComment() {
-    const { text, ownerName } = this.props
+    const { text, ownerName, viewportIsMobile } = this.props
 
     return (
-      <div style={styles.comment}>
+      <StyleRoot style={styles.comment}>
         <Marger bottom="1">
           <Text color="font1" size="tiny" weight="regular">
             {ownerName}
           </Text>
         </Marger>
-        <Marger top="1">{text}</Marger>
+        <Marger top="1">
+          <Text
+            color="font1"
+            size={viewportIsMobile ? 'tiny' : 'default'}
+            weight="light"
+          >
+            {text}
+          </Text>
+        </Marger>
         <span style={styles.comment.arrow} />
-      </div>
+      </StyleRoot>
     )
   }
 }
@@ -56,29 +65,35 @@ const styles = {
   },
   comment: {
     position: 'relative',
-    borderWidth: '2px',
+    borderWidth: 2,
     backgroundColor: COLORS.background3,
     borderColor: COLORS.background3,
     color: COLORS.font1,
-    padding: '30px',
-    fontSize: '16px',
+    padding: 30,
+    fontSize: 16,
     marginLeft: 20,
-
     [`@media (min-width: ${ScreenConfig['S'].min}px)`]: {
       marginLeft: 35,
     },
 
     arrow: {
       position: 'absolute',
-      top: '30px',
+      top: 20,
       display: 'block',
       width: 0,
       height: 0,
-      borderWidth: '10px',
+      borderWidth: 10,
       borderStyle: 'solid',
       borderColor: 'transparent',
       borderRightColor: COLORS.background3,
-      left: '-20px',
+      left: -20,
+      [`@media (min-width: ${ScreenConfig['S'].min}px)`]: {
+        top: 35,
+      },
     },
   },
 }
+
+export const Comment = mediaQueries(Radium(CommentComponent), {
+  viewportIsMobile: true,
+})